Add roofline analyze test coverage
